requirements.txt
pyqt5
groq
python-dotenv
pillow